Use vitest for tests
